Move .evaluateFormulas() to lib/jellyscript
