+++ lib/actions/index.js
@@ -21,6 +21,7 @@ const EventEmitter = require('events').EventEmitter
 const objectTemplate = require('object-template')
 const library = require('./library')
 const time = require('./time')
+const jellyscript = require('../jellyscript')
 
 class ActionRequestWorker {
 	constructor (jellyfish, session) {
@@ -214,7 +215,7 @@ class ActionRequestWorker {
 				required: argumentNames
 			}
 
-		const computedArguments = this.jellyfish.evaluateFormulas(argumentsSchema, args)
+		const computedArguments = jellyscript.evaluateObject(argumentsSchema, args)
 		if (!this.jellyfish.matchesSchema(argumentsSchema, computedArguments)) {
 			throw new this.jellyfish.errors.JellyfishSchemaMismatch('Arguments do not match')
 		}
