+++ lib/core/json-schema.js
@@ -20,9 +20,6 @@ const debug = require('debug')('jellyfish:json-schema')
 const mergeSchema = require('json-schema-merge-allof')
 const ajvKeywords = require('ajv-keywords')
 const errors = require('./errors')
-const utils = require('./utils')
-
-const FORMULA_PROPERTY = '$formula'
 
 const JSON_SCHEMA_SCHEMA = {
 	$schema: 'http://json-schema.org/draft-07/schema#',
@@ -228,10 +225,6 @@ const JSON_SCHEMA_SCHEMA = {
 					uniqueItems: true
 				}
 			]
-		},
-		[FORMULA_PROPERTY]: {
-			type: 'string',
-			pattern: '^[a-zA-Z0-9()$,\\s]+$'
 		},
 		format: {
 			type: 'string'
@@ -477,68 +470,6 @@ const disallowAdditionalProperties = (schema, options) => {
 
 	schema.properties = _.mapValues(schema.properties, disallowAdditionalProperties)
 	return schema
-}
-
-const eachDeep = (object, callback, breadcrumb = []) => {
-	for (const key of Object.keys(object)) {
-		const value = object[key]
-		const absoluteKey = breadcrumb.concat([ key ])
-
-		if (_.isPlainObject(value)) {
-			eachDeep(value, callback, absoluteKey)
-			continue
-		}
-
-		if (_.isArray(value)) {
-			_.each(value, (element, index) => {
-				eachDeep(element, callback, absoluteKey.concat([ index ]))
-			})
-
-			continue
-		}
-
-		// eslint-disable-next-line callback-return
-		callback(value, absoluteKey)
-	}
-}
-
-const getRealObjectPath = (schemaPath) => {
-	return schemaPath.slice(0, schemaPath.length - 1).filter((fragment) => {
-		if (_.isNumber(fragment)) {
-			return false
-		}
-
-		return ![
-			'properties',
-			'anyOf',
-			'allOf',
-			'oneOf'
-		].includes(fragment)
-	})
-}
-
-const formulaSchemaCache = new Map()
-
-exports.getFormulasPaths = (schema) => {
-	const hash = utils.hashObject(schema)
-
-	if (formulaSchemaCache.has(hash)) {
-		return formulaSchemaCache.get(hash)
-	}
-
-	const paths = []
-
-	eachDeep(schema, (value, key) => {
-		if (_.last(key) === FORMULA_PROPERTY) {
-			paths.push({
-				formula: value,
-				output: getRealObjectPath(key)
-			})
-		}
-	})
-
-	formulaSchemaCache.set(hash, paths)
-	return paths
 }
 
 /**
