+++ lib/core/kernel.js
@@ -240,7 +240,7 @@ module.exports = class Kernel {
 			throw new this.errors.JellyfishUnknownCardType(`Unknown type: ${card.type}`)
 		}
 
-		const processedCard = this.evaluateFormulas(schema, card)
+		const processedCard = jellyscript.evaluateObject(schema, card)
 		jsonSchema.validate(schema, processedCard)
 
 		const bucket = this.getBucketForType(processedCard.type)
@@ -250,55 +250,6 @@ module.exports = class Kernel {
 		}
 
 		return this.backend.insertElement(bucket, processedCard)
-	}
-
-	/**
-   * @summary Evaluate schema formulas against an object
-   * @function
-   * @public
-   *
-	 * @param {Object} schema - schema
-   * @param {Object} object - input object
-   * @returns {Object} evaluated object
-   *
-   * @example
-   * const kernel = new Kernel(backend, { ... })
-   * await kernel.initialize()
-   *
-	 * const result = kernel.evaluateFormulas({
-	 *   type: 'object',
-	 *   properties: {
-	 *     foo: {
-	 *       type: 'number',
-	 *       $formula: 'MAX(input, 2)'
-	 *     }
-	 *   }
-	 * }, {
-	 *   foo: 1
-	 * })
-	 *
-	 * console.log(result.foo)
-	 * > 2
-   */
-	// eslint-disable-next-line class-methods-use-this
-	evaluateFormulas (schema, object) {
-		for (const path of jsonSchema.getFormulasPaths(schema)) {
-			if (_.isEmpty(object)) {
-				continue
-			}
-
-			const result = jellyscript.evaluate(path.formula, {
-				context: object,
-				input: _.get(object, path.output)
-			})
-
-			if (!_.isNull(result.value)) {
-				// Mutates input object
-				_.set(object, path.output, result.value)
-			}
-		}
-
-		return object
 	}
 
 	/**
