+++ lib/core/utils.js
@@ -16,29 +16,6 @@
 
 const _ = require('lodash')
 const EventEmitter = require('events').EventEmitter
-const objectHash = require('object-hash')
-
-/**
- * @summary Hash a JavaScript object
- * @function
- * @public
- *
- * @param {Object} object - object
- * @returns {String} object hash
- *
- * @example
- * const string = utils.hashObject({ foo: 'bar' })
- * console.log(string)
- */
-exports.hashObject = (object) => {
-	return objectHash(object, {
-		algorithm: 'sha1',
-		ignoreUnknown: true,
-
-		// This in particular is a HUGE improvement
-		respectType: false
-	})
-}
 
 exports.mergeStreams = async (keys, createFunction) => {
 	const emitter = new EventEmitter()
