+++ lib/jellyscript/index.js
@@ -19,6 +19,7 @@ const _ = require('lodash')
 const formula = require('formulajs')
 const staticEval = require('static-eval')
 const esprima = require('esprima')
+const card = require('./card')
 
 // Based on https://stackoverflow.com/a/17201493/1641422
 const DEFAULT_ITERATIONS = 10000
@@ -128,4 +129,24 @@ exports.evaluate = (expression, options) => {
 		value: result,
 		watchers: []
 	}
+}
+
+exports.evaluateObject = (schema, object) => {
+	for (const path of card.getFormulasPaths(schema)) {
+		if (_.isEmpty(object)) {
+			continue
+		}
+
+		const result = exports.evaluate(path.formula, {
+			context: object,
+			input: _.get(object, path.output)
+		})
+
+		if (!_.isNull(result.value)) {
+			// Mutates input object
+			_.set(object, path.output, result.value)
+		}
+	}
+
+	return object
 }
