+++ test/core/json-schema.spec.js
@@ -322,108 +322,3 @@ _.each(MERGE_TEST_CASES, (testCase, index) => {
 		}
 	})
 })
-
-ava.test('.getFormulasPaths() should return an empty array given no formulas', (test) => {
-	const paths = jsonSchema.getFormulasPaths({
-		type: 'object',
-		properties: {
-			foo: {
-				type: 'string'
-			},
-			bar: {
-				type: 'string'
-			}
-		}
-	})
-
-	test.deepEqual(paths, [])
-})
-
-ava.test('.getFormulasPaths() should return one property with formulas', (test) => {
-	const paths = jsonSchema.getFormulasPaths({
-		type: 'object',
-		properties: {
-			foo: {
-				type: 'string',
-				$formula: 'UPPER(input)'
-			},
-			bar: {
-				type: 'string'
-			}
-		}
-	})
-
-	test.deepEqual(paths, [
-		{
-			formula: 'UPPER(input)',
-			output: [ 'foo' ]
-		}
-	])
-})
-
-ava.test('.getFormulasPaths() should return nested properties with formulas', (test) => {
-	const paths = jsonSchema.getFormulasPaths({
-		type: 'object',
-		properties: {
-			foo: {
-				type: 'string',
-				$formula: 'UPPER(input)'
-			},
-			bar: {
-				type: 'object',
-				properties: {
-					baz: {
-						type: 'number',
-						$formula: 'POW(input, 2)'
-					}
-				}
-			}
-		}
-	})
-
-	test.deepEqual(paths, [
-		{
-			formula: 'UPPER(input)',
-			output: [ 'foo' ]
-		},
-		{
-			formula: 'POW(input, 2)',
-			output: [ 'bar', 'baz' ]
-		}
-	])
-})
-
-ava.test('.getFormulasPaths() should return properties inside arrays', (test) => {
-	const paths = jsonSchema.getFormulasPaths({
-		type: 'object',
-		anyOf: [
-			{
-				properties: {
-					foo: {
-						type: 'string',
-						$formula: 'UPPER(input)'
-					}
-				}
-			},
-			{
-				properties: {
-					bar: {
-						type: 'string',
-						$formula: 'LOWER(input)'
-					}
-				}
-			}
-		]
-	})
-
-	test.deepEqual(paths, [
-		{
-			formula: 'UPPER(input)',
-			output: [ 'foo' ]
-		},
-		{
-			formula: 'LOWER(input)',
-			output: [ 'bar' ]
-		}
-	])
-})
